fix: fixed the dead lock issue of random_start_ver
This commit fixed the dead lock issue caused by the start &&
(Q == 9'b10) && done in the if statement. The reason of the dead
lock was that it needed done to set the newstart to 1, but the done
would always be 0 if the newstart haden't been 1. Now, the state of
the inner random_assign is recorded with a reg random_assign_busy.

This commit also perfed the module by deleting the counter inside it.
Since the module was controlled by the start only, it did not need
the counter.